Review-fixes: /producten zonder zoekterm, rooktest, robuustere worker
- catalog-search: ORDER BY 0 was een kolomnummer => 500 op /producten zonder q
- scripts/smoke.ts (npm run smoke): 25 pagina/API-controles
- productpagina: willekeurige categorie-suggesties weg (traag, irrelevant)
- worker: signalen loggen, dubbele Ctrl+C binnen 10 s negeren
- sitemap: revalidate 1 uur

+++ lib/catalog-search.ts
@@ -98,7 +98,7 @@ export async function searchGroups(
         ws.map((t) => Prisma.sql`lower(cp2.name) ~ ${"(^| )" + t.replace(/[^a-z0-9]/g, "") + "( |$)"}`),
         " AND ",
       )} THEN 0 ELSE 1 END`
-    : Prisma.sql`0`;
+    : Prisma.sql`0::int`; // niet kaal "0": ORDER BY 0 is een kolomnummer
   const order =
     o.sort === "price"
       ? Prisma.sql`g.min_cents ASC, g.stores DESC`
